Unify analyze + diff into one processor and one route contract
analyzeInProcess and diffInProcess ran the same analyzer pipeline but
lived in two separate wrappers; the dashboard path did too (POST
/analyze returned 202 + streamed progress via sockets, POST /diff-check
was synchronous and emitted nothing); the CLI did too (runAnalyze used
StepTracker, runAnalyzeDiff a plain spinner). Every analyze UX
improvement drifted off the diff path — the pre-flight LLM prompt gap
fixed earlier in this branch was the most recent example; the missing
progress panel on diff is the other symptom.

Changes:

- New `analyze-core.ts` — the single mode-agnostic computation path
  (lock → load LATEST baseline → runAnalysis w/ skipStash from mode →
  buildGraph → runViolationPipeline → drain usage → enforce location
  invariant). Returns AnalyzeCoreResult with everything both modes
  need. No disk IO beyond the analyze lock.
- New `analyze-persist.ts` — persistFullAnalysis (writes analyses/
  LATEST/history, deletes diff.json, updates lastAnalyzed) and
  persistDiffAnalysis (hydrates resolved rows from baseline, computes
  affectedNodeIds, writes diff.json).
- `analyzeInProcess` / `diffInProcess` become two-line wrappers over
  analyzeCore + the matching persister. Public shapes and subpath
  exports unchanged.
- `POST /api/repos/:id/diff-check` now returns 202, registers with
  analysis-registry, runs in the background with a socket tracker
  and createSocketLlmEstimateHandler, and emits
  analysis:progress/llm-estimate/violations:ready/analysis:complete —
  same event stream full analyze uses. `POST /analyze/cancel` works
  for diffs through the same registry slot.
- `useDiffCheck` rewritten: POST, wait for analysis:complete over
  sockets (or analysis:canceled / error via progress), then GET
  diff.json. Same {diffResult, isChecking, error, run, load, clear}
  surface so RepoGraphPage only needed to pass `onEvent` through.
- CLI `runAnalyzeDiff` swaps the p.spinner() for StepTracker +
  renderSteps + renderPhase — identical checklist UX to runAnalyze.
- DiffSnapshot grows `usage: UsageRecord[]` (optional, backwards
  compatible). Diff runs that hit LLM were previously discarding the
  usage — now we record it for cost-tracking.
- Concurrency tightened — the single analyze lock now guards
  analyze-vs-diff too, not just analyze-vs-analyze.

Behavior preserved (diff.json keeps all its fields, including the
hydrated resolvedViolations and name-keyed affectedNodeIds). Disk
shapes unchanged except for the additive optional `usage` field.
